Fix null return type in get_last_batch_id and docblock spacing
- Cast get_last_batch_id() return to int to prevent TypeError when
  feed_last_queued_item_id is not yet set in persistent storage
- Fix excessive spaces in next_scheduled_action docblock to satisfy
  Squiz.Commenting.FunctionComment.SpacingAfterParamType

+++ src/FeedGenerator.php
@@ -678,7 +678,7 @@ class FeedGenerator extends AbstractChainedJob {
 			Pinterest_For_Woocommerce::save_data( 'feed_last_queued_item_id', 0 );
 		}
 		// Get last fetched ID to start from the next item after it.
-		return Pinterest_For_Woocommerce::get_data( 'feed_last_queued_item_id' );
+		return (int) Pinterest_For_Woocommerce::get_data( 'feed_last_queued_item_id' );
 	}
 
 	/**
